Update prebuilt.py to sync to latest revision before pushing.
This fixes a bug where prebuilt.py is unable to push if it is using
manifest-versions to sync to an old revision. This bug occurs because
manifest-versions is pointing at an old version of the repository so
the prebuilt pusher is never able to push.

BUG=chrome-os-partner:3959
TEST=Test with --dry-run push. Also run unit test suite.

Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: David James <davidjames@chromium.org>
(cherry picked from commit 1b6e67aef79696d74bcee3d2fce252c8397e592a)

Change-Id: I3e254586c50323d561890c90df81b82bd52987c4
Reviewed-on: http://gerrit.chromium.org/gerrit/1241
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Nick Sanders <nsanders@chromium.org>