Fix bug in manifest_version where we fail to push the version file but leave state around.
The problem here is we correctly retry, but since we removed the code to abandon the change,
we still have the local commit in our tree -> we use a manifest that has this local commit
from before. This is bad. Very bad.

BUG=chromium-os:15998 chromeos-partner:4316
TEST=Ran mario-release with --debug through to make_chroot and ran unittests

Reviewed-on: http://gerrit.chromium.org/gerrit/1898
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>

(cherry picked from commit 75122c88adc1d285edebca08cc5afa99c1843f70)

Change-Id: I8fdac77d9c6d72c1346fa26641c39aa7d27ff8d8
Reviewed-on: http://gerrit.chromium.org/gerrit/1995
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>