Cleanup uprev code, centralize hardcoded assumptions about the remote
Specifically, move away from pushinsteadof (which is just sed'ing and
hardcodes assumptions about the fetch url), instead using the available
repo provided info to define a proper pushurl; in the process remove any
pushinsteadof's that were already created removing the need for a clobber.

Additionally, gut dead code from the uprev function itself that has no
actual affect.  This leaves behind a forced reset precluding the
updated uprev configs for the sake of avoiding potential breakage-
that said it's very likely to be unneeded.

Note that the repo forall is intended to be removed down the line;
that's tracked in chromium-os:24503.

Finally, this change is the exact same as Ic049086d, with a single
character of difference; &> /dev/null becomes 2> /dev/null.

This is required because the builders have /bin/sh -> dash, and dash
is stupid and treats '&> /dev/null' as '> /dev/null &'.  zsh, bash,
most sane shells don't do that, but old standards are such that you
should use '> /dev/null 2>&1' (in that order should the shell be
stupid about it's dup2 call) instead.

That insanity aside, we only want to silence complaints if the section
doesn't exist, thus we silence stderr alone.

BUG=chromium-os:24413, chromium:108196
TEST=no staging buildbots exist to test against.  Testeded via manually
     walking through the steps.

Change-Id: Ic049a86d98b2bc7f170357118fba8efdc128cfcc
Reviewed-on: https://gerrit.chromium.org/gerrit/13596
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>