Output version incrementing via info, and assert the results.
Current code is quiet when it does the bump (due to being debug
output), nor does it actually ensure we just got a different
version back.

That missing latter check means that if somehow that
step fails (say grabbing from http vs us getting the value
from ssh), we'll spot it.  Unlikely, but since it's mildly
critical, leving an assert.

BUG=chromium-os:27089
TEST=None

Change-Id: Ic1f4168b8ebabcd6d2a8d3fd7399c8d1d442e134
Reviewed-on: https://gerrit.chromium.org/gerrit/24017
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>