Use repo init -b for branch switching.
Now that repo is upgraded to 1.9.4, re-init no longer needs to delete
the 'default' branch - repo does this by itself.  There are some
known oddities with this support as of v1.9.4- that said there is
some known breakages from the previous approach; see:
https://groups.google.com/forum/?fromgroups#!msg/repo-discuss/4WmUJ2ttN8o/ssYVLO5TCVYJ
for examples.

That's being tracked upstream and the outstanding issues are suppressed
via this patch.

Additionally, the cbuildbot/repo interaction that leads to a
detached HEAD (which *cannot* be replicated via repo directly,
which means it *definitely* is something cbuildbot is doing-
likely a bad interaction of the previous --detach approach during
transitioning/patch application) is detected, warned, and suppressed.

Further investigation into that will occur if/when required; strongly
expect the issue to go away after the --detach code is removed however.

BUG=chromium-os:31912, chromium-os:32054
TEST=remote trybot switching branches.

Change-Id: I3f20a8af5507b17e98d86d7e8ca8d34ecbda5c0d
Reviewed-on: https://gerrit.chromium.org/gerrit/25450
Tested-by: Brian Harring <ferringb@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>