manifest_version: allow instantiation via a source repo
A common workflow is to say "give me the version of this source repo".
As such, rather than forcing people to give the full path to the version
(which is already a buildbot constant), allow them to give the root of
the source repo.  We'll take care of the rest.

BUG=chromium-os:19140
TEST=`./buildbot/run_tests` passed

Change-Id: Id1698109e0c5e1fc85178bd4e522c58a4600aa54
Reviewed-on: https://gerrit.chromium.org/gerrit/33628
Reviewed-by: Brian Harring <ferringb@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35385