Revert "speed up firmware builders"
This reverts commit a3fb8373577ad85dd6592c28e66348eb4519c552

The virtual/chromeos-firmware package apparently doesn't pull
in all the firmware packages that I expected.  Until we can
sort that out, we need to go back to a full build so that all
the packages needed get installed.  We'll sort out the policy
in chromium-os:19140 and then re-land this CL.

Change-Id: Iaa164b6bbde34a65fa3247bffed86dd7b2008af0
Reviewed-on: https://gerrit.chromium.org/gerrit/35522
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>