Force gs.py to use generation rather than sequence-match.
Currently no consuming code has been found sans unittests so
this is reasonably safe.  For existing usage where it was
just a fail-if-exists check, the change restores functionality;
for any sequence tracking code (which I've yet to find sans locks
that don't use this pathway) it will break; it was broken via
the referenced bug anyways, thus is decided to move forward with.

BUG=chromium-os:37718
TEST=validation of the header, but nothing beyond that; TBR.

Change-Id: Ic1d42b0e9dca1be60105f86aa0eaaed6020e0f40
Reviewed-on: https://gerrit.chromium.org/gerrit/40933
Reviewed-by: Brian Harring <ferringb@chromium.org>
Tested-by: Brian Harring <ferringb@chromium.org>