crosfw: Enable building beaglebone and VBOOT debugging
The beaglebone has underscore in its name, so is not currently supported. The
fix is to only perform the Chrome OS underscore checking when we know it
is a Chrome OS board.

Also we should always enable VBOOT debugging, so correct that.

BUG=none
TEST=manual
$ crosfw -b am335x_evm -w
See that it now works correctly.
The VBOOT change was tested with a special U-Boot, CLs are still pending in
gerrit.

Change-Id: I19a73b4b0ed617c6a6384bd424858cac38fac374
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61501
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>