Add script for creating projects.
The purpose of this script is to ensure that certain often overlooked
steps (specifying the parent for example) is done- in this case
automatically calculated if not specified- while specifying other
defaults.

This basically is a defaulted version of invoking gerrit
create-project, and should be the preferred norm going forward.

BUG=chromium:138826
TEST=Run the script in dry-run mode and verify the command line.

Change-Id: I2e7cc0aca5a68b0bfd7d7f55d7d0bce70961abb9
Reviewed-on: https://gerrit.chromium.org/gerrit/62140
Commit-Queue: David James <davidjames@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>