Don't try to clean up the tee process.
We don't want the critical section to try to clean up the tee process,
so we run Tee (forked off) outside of it. This prevents a deadlock
because the Tee process only exits when its pipe is closed, and the
critical section accidentally holds on to that file handle.

BUG=chromium:261402
TEST=Test on buildbot and verify hang is gone.
TEST=Paladin trybot run.

Change-Id: Ia457bbc9fe8c60de6b5c7474b5f490776a2cac41
Reviewed-on: https://gerrit.chromium.org/gerrit/62427
Commit-Queue: David James <davidjames@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>