Don't set -chrome_debug on perf builder.
-chrome_debug disables the generation of debug symbols. We need debug
symbols in production so that they can be stripped and uploaded to the
crash server. We should do the same in perf builds to have more
representative builds, and also to allow developers to debug any
performance problems that are found in the tests.

BUG=none
TEST=All unit tests.

Change-Id: I5e15ef0ac32d89c6ffd39ab1d8c2a1ce17d8f168
Reviewed-on: https://gerrit.chromium.org/gerrit/62923
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>