Add generate_delta_sysroot.py
This script generates a tarball containing a sysroot that can be patched
over extracted prebuilt package contents to create a complete sysroot.

Usage:
  generate_delta_sysroot --board=<board> --out-dir=<out_dir>
  [--out-file=<out_file> --out-batch=<batch>]

  Creates a delta sysroot tarball of name <out_file> in <out_dir>
  containing batch file <batch> for board <board>

The script works by:
  - Emerges all packages for a board into a build sysroot directory.
  - Unpack all packages for a board into a temp sysroot directory.
  - Creates a batch file using:
    rsync -rplgoDc --delete --write-batch=<batch> <build_sys> <tmp_sys>
  - Puts the batch file inside a tarball.

This batch file can be applied to any directory identical to the
temporary folder. This can be achieved by using build_packages with an
--unpackonly flag. Then, in order to get a complete sysroot, it's sufficient
to use:

Note that the script assumes that there exists a built sysroot already.
In fact, the packages will be fetched from that sysroot.

rsync -rplgoDc --delete --read-batch=<batch> <extract_pkgs_dir>

BUG=None
TEST=scripts/generate_delta_sysroot_unittest.py

Change-Id: I22dafd8ba900cb48956a32ff87d3d1206f14d322
Signed-off-by: Thiago Goncales <thiagog@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56653
Reviewed-by: Ryan Cui <rcui@chromium.org>