commandline: sort log levels
The sorting of this constant doesn't matter to the code, but it does
get used in the --help output.  So let's sort the var so it indicates
the relative levels to users.  Otherwise they need to look up the
logging documentation to figure out which is higher priority.

BUG=None
TEST=`upload_symbols --help` shows log levels sorted correctly

Change-Id: Idcb34472462b3726c1cdbf95198c151de662b1c6
Reviewed-on: https://gerrit.chromium.org/gerrit/65532
Reviewed-by: Marc MERLIN <merlin@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>