Make the commit queue red when it fails.
Right now, the CommitQueueCompletion stage is a ForgivingStage, meaning
that it ignores all errors. This means that exceptions in this stage
are completely ignored, and builders that fail this stage do not turn
red simply because this stage failed.

Besides the master, the completion stages also run on slaves, to report
their status back to the master. If the build failed, they throw an
exception here. For slave builders, marking this stage 'red' would be
redundant, since the build itself would already be red. This is the only
case where we still report a warning. In all other cases, we report an
error when an exception is thrown.

BUG=chromium:288227
TEST=Run mario-paladin and lumpy-paladin in debug mode with an example
failure and example success. Also confirm that failures other than
ImportantBuilderFailedException are treated as failures.

Change-Id: I6bc9c85ed3e65a193739e02576a80d0f38f91095
Reviewed-on: https://chromium-review.googlesource.com/168760
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>