Use borg mirror for manifest-versions.
Manifest-versions is currently using gerrit which can not only be very
slow but can also have a high rate of downtime. Since this is a very
large repository and we spend a lot of time fetching it, switching to a
GoB mirror from gerrit would help improve the speed and reliability of
our builds.

BUG=chromium:290450
TEST=Example commit queue and release runs.

Change-Id: I72fb2456bd35d8e3c8160326ecc8846ddcdc0f85
Reviewed-on: https://chromium-review.googlesource.com/169204
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>