autotest_quickmerge: programatically detect portage downgrade list
Instead of having a hard-coded list of autotest related packages to be
downgraded by autotest-quickmerge, this CL causes autotest_quickmerge to
parse the per-ebuild autotest packaging log files in the sysroot to
determine which packages to downgrade.

BUG=chromium:289181
TEST=Compared full directory print of sysroot autotest between a fresh
build_packages and (autotest_quickmerge && build_packages). Saw that all
but 4 package files were being correctly restored. I believe these
packages are being provided by an ebuild that has the same ${PV} name as
another ebuild, and so its log file is getting overwritten and hence it
is not being added the the downgrade list; a fix for that in the
autotest.eclass is forthcoming.

Change-Id: Id7b4b2b03b575528b9437a464cab8eb1eff1c166
Reviewed-on: https://chromium-review.googlesource.com/169635
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>