Don't actually mess with CLs in remote trybot runs of PreCQLauncher.
For initial testing, PreCQLauncher was set up so that it could be
run as a remote trybot and would affect production, even if it did
not have the --buildbot flag. For now, though, if someone runs it
as a --remote run, it is better to only affect production if they
pass in the --buildbot flag.

BUG=none
TEST=Run PreCQLauncher and verify it does not mess with any CLs.

Change-Id: Ifaa321e5465e5bd931b0e35a59dfd9e3e2afd6e1
Reviewed-on: https://chromium-review.googlesource.com/170166
Reviewed-by: Stefan Zager <szager@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>