Don't crash when we get a long patch series.
Limit the number of CLs to include in the name of the tryjob at 10, so
that we don't get a validation error for a long name. This helps us run
more CLs at once.

There is still a limitation on the number of CLs that can be run
(extra_args is limited to 1024 characters), but we're much less likely
to hit that. We can handle around 60 patches now.

BUG=chromium:258461
TEST=Run a big list of CLs through the tryserver.

Change-Id: I7883a3ea9357d7804598b2904d437b9d0d66bf8a
Reviewed-on: https://chromium-review.googlesource.com/170261
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>