cros_build_lib.ContextManagerStack: Comments and minor bug fix.
This change mostly adds clarifying comments to the ContextManagerStack
class.  It also fixes two minor bugs in the code that were probably
never triggered:

1) If a context manager object was None it was still adding it to
self._stack, which would result in a failure during __exit__.  Now it
skips the None object.

2) At the end of __exit__ self._stack was only reset to [] if there was
still an exception to raise.  Now it always resets it.  This only
matters if the same ContextManagerStack object is used twice, which we
never do.

BUG=None
TEST=pylint
TEST=unittest

Change-Id: If60df158af189af808779ca26b2df2424d2a0376
Reviewed-on: https://chromium-review.googlesource.com/170488
Tested-by: Matt Tennant <mtennant@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Matt Tennant <mtennant@chromium.org>