Re-request gerrit query results to avoid cache staleness.
Change data returned by the /changes?q=query gerrit endpoint can be
stale, depending on the status of the gerrit cache.  This patch
causes all query results to be re-requested from gerrit using an
endpoint that doesn't go through the cache.

This is an update of CL:171755, which broke the CQ.  I added a test
to the new gerrit_unittest.py, which is still in review at CL:170475.

BUG=302072
TEST=lib/gerrit_unittest.py

Change-Id: Ia2dbfb4e39f0fae40a0108ee0104723076284f77
Reviewed-on: https://chromium-review.googlesource.com/172256
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Tested-by: Stefan Zager <szager@chromium.org>