Include slave build statuses in master build metadata.json
This CL adds information about slave builds to the metadata.json
file of a CQ-master build. These are added as an entry of the form:

{
...
  'slave_targets' : {
    'x86-mario-paladin' : {
      status : 'fail',
      message : 'string-formatted message'
    },
    'lumpy-paladin' : {
      ...
    },
    ...
  }
...
}

Adding this slave metadata required a way for the ReportStage to reference
to CommitQueueCompletion stage, in order to pull in the slave statuses,
and a way to flatten a BuilderStatus object into a json-encodable string-
to-string dictionary.

BUG=chromium:287880
TEST=Ran a number of local `cbuildbot --local --buildbot --debug` runs,
and examined the dropped metadata.json file, which contained a
'slave_targets' key. In my local tests, I only saw slave results for
the build configuration that I kicked off; in prod this should also
include the other CQ build slaves. I'm convinced the easiest way to test
this is to just put this change in.

Change-Id: I5aeb5d9d470e815ae7e7e7fbf75bad7a2f286cf9
Reviewed-on: https://chromium-review.googlesource.com/172140
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Matt Tennant <mtennant@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>