Track how often a patch has failed in metadata.json.
This CL adds a count of how often a patch has failed to metadata.json.

"fail" is the number of times the latest version of the patch has failed.
"total_fail" is the number of times all versions of the patch have failed.
"pass" and "total_pass" are included similarly.

This data can be used to estimate how many patches could have busted the CQ.
(If total_fail > fail, then the earlier version of the patch could've busted
 the CQ.)

BUG=chromium:308640
TEST=Run remote trybot and verify metadata.json includes new fields.
TEST=unit tests
Change-Id: I849145f399ad31e58bb204a115b0db7a160c16f4
Reviewed-on: https://chromium-review.googlesource.com/173583
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>