Publish ebuild uprevs even if the build failed.
Currently we only publish ebuild uprevs if the build succeeded, the completion
stage succeeded and all changes were submitted. In the case where the
CQ submitted some but not all of the changes, or where the tree has chumped
changes, the tree can be left in an inconsistent state. In this case,
sync to a clean tree and uprev any changes that are already in the tree.

BUG=chromium:310671, chromium:309800, chromium:310945, chromium:311334
TEST=--buildbot --debug run with failing stages

Change-Id: I1cfbf5eca872d3dae0176ff891242ff3f2d876b8
Reviewed-on: https://chromium-review.googlesource.com/174543
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>