Enable HWTest phase for all R35 canaries that might need it.
As of a recent change, it is now safe to request HWTest for boards
that don't exist.  This is done so that builders can request HWTest
without regard to whether hardware is available, so that the
builders don't have to track the inventory in the lab.

This change adds a HWTest phase to all canaries that are expected
to eventually need testing for R35, regardless of whether hardware
currently exists in the lab.

BUG=chromium:353541
TEST=python cbuildbot_config_unittest.py

Change-Id: I8327551cda9eabc72064b800b4172ee3f13b8883
Reviewed-on: https://chromium-review.googlesource.com/196610
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Yu-Ju Hong <yjhong@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>