overlays: build list by masters
Rather than hardcode the list of overlays that we always use, look at
the masters list in the layout.conf and do so recursively.

We retain the hardcoded legacy list for now, but it'll be removed in
a follow up commit once this is stable.

BUG=chromium:407604
TEST=`./cbuildbot/run_tests` passes
TEST=`cbuildbot amd64-generic-full` passes (a public manifest)
TEST=`cbuildbot pre-cq-group` passes (lots of internal manifests)

Previous-Reviewed-on: https://chromium-review.googlesource.com/220191
(cherry picked from commit 658ec68bb63f5797f9ada33e93bde48f04b04f09)

Change-Id: I7e96c679a605008da3db6d95fe263e5a28d617bc
Reviewed-on: https://chromium-review.googlesource.com/233901
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>