Disable unittests on veyron_pinky factory pre-flight
It looks like veyron_pinky preflight is dying on the unittest phase.
This does not appear to run on the canary, so this failure was not
noticed before the branch point.

This build step is skipped in many other places, particularly when
using new and different achitectures, so I think we can get by disabling
it here on the factory branch.

BUG=chromium:443460
TEST=None

Change-Id: I734c1d02f06c9c239d074217c7a1d3f18422fb5c
Reviewed-on: https://chromium-review.googlesource.com/237074
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>