RunCurl: run with --fail by default
In order to get the extended exit code failures for HTTP 4xx and such, we
have to run curl with the --fail option.  It's easy for people to forget
to include this though, so make it an explicit keyword to RunCurl and set
it on by default.

This lets lib.cache abort quickly when given a URL that is 404.

BUG=chromium:478224
TEST=`./cbuildbot/run_tests` passes

Change-Id: I0d0d09828be950536174d9f77c1e4011679578f6
(cherry picked from commit bf6b36ce3d24bcfda2e4c78addb576752648693e)
Reviewed-on: https://chromium-review.googlesource.com/1139653
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>