Avoid generating AFDO profiles when Chrome was not uprevved.
The *-chrome-pfq builders stop as soon as they detect chrome
was not uprevved. However, the pre-fligh-release builders dont
stop when Chrome has not been uprevved and, therefore, AFDO
profile collection will be done although it is not necessary.

This CL changes the implementation of the fix for this bug.
Instead of generating profiles only when chrome was uprevved,
we generate profiles for a version of chrome only when it
has not already been generated.

This version of the fix has 2 advantages:
- should be more resilient to failures. In the previous fix,
it was possible that the generation of the profile failed (because
of hardware lab issues, for example) in which case the profile
would not have been generated until the next chrome uprev. This
would have been ok on the branches where chrome is uprevved frequently
like ToT but it would not have been good on release branches where
chrome is not uprevved everyday.
- It avoids a corner case where it was possible the process would have
complained about stale profiles if chrome was not uprevved in more
than 7 days. With the new fix, the file used to detect how stale
the profile is, is updated everytime we detect a new profile is not
needed. So, with the new fix, the process will only complain if
a new version of chrome does not have a generated profile in more
than 7 days.

BUG=chromium:472895
TEST=cbuildbot lumpy-pre-flight (done)
     cbuildbot lumpy-chrome-pfq (done)
     cbuildbot lumpy-release (done)
     run_tests

Reviewed-on: https://chromium-review.googlesource.com/285542
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Dharani Govindan <dharani@chromium.org>
(cherry picked from commit 091213492e27609f64d25392a58a46bae97351d0)

Change-Id: Ic159657ffcd085034ac2452ae036d82c7bab4d5e
Reviewed-on: https://chromium-review.googlesource.com/287614
Trybot-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>