Use different pool for AFDO profile collection
We are having several issues with the "suites" pool. So lets
change the AFDO profile collection to use the BVT pool and change
the priority of the test to make sure it is executed soon enough.

Adding this change to the R45 branch. Change on ToT is working fine.
The R45 branch does not have the change and it is failing constantly.
So put the change on the branch.

BUG=None
TEST=run_tests
     cbuildbot lumpy-chrome-pfq

Change-Id: I61d082eaa87cd40adf87c50e594ac29b8842dd1e
Previous-Reviewed-on: https://chromium-review.googlesource.com/287850
(cherry picked from commit b37eac0e2ede1e8bb4849f8f2e7179d3c5ec780c)
Reviewed-on: https://chromium-review.googlesource.com/289930
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Trybot-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>