[re-land] sync_stages: fix _VerifyMasterId to respect milestone_version
This is a re-land of https://chromium-review.googlesource.com/344762

The original relied on cidb to return a build status is that it was not
returning. The issue was masked in unit tests, because fake_cidb *was*
returning this key. In this patch, the underlying issue is fixed, and
fake_cidb is brought into compliance with cidb in terms of exactly what
keys it returns (so that in the future, unit tests will not be fooled).

BUG=chromium:611929
TEST=cidb_integration_test; unit tests

Change-Id: Ib002348e957f75cc12fc626c2729e0c27e0bb2a7
Previous-Reviewed-on: https://chromium-review.googlesource.com/345430
(cherry picked from commit 2102e3b3f10ecf50e4d1a216d9fed222aa360201)
Reviewed-on: https://chromium-review.googlesource.com/345973
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>