AFDO: increase allowed stale for profile on inactive branch.
R52 is mostly inactive. Chrome is not being uprevved anymore.
But the AFDO process is trying to check if the AFDO profile is
not stale and it is failing because it has not been updated in
the last 7 days. To workaround this, increase the stale limit
to 35 days which should be more than enough.

BUG=chromium:637307
TEST=none

Change-Id: I3212eaf89c1328113f751d243fae3aece6ccf869
Reviewed-on: https://chromium-review.googlesource.com/373683
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Trybot-Ready: Luis Lozano <llozano@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>