gs: downgrade to gsutil 4.19
Ever since the 4.20 update the canary builders on GCE have been
dying in AUTest, failing to find artifacts in a particular GS bucket.

The release branch builders, which still have 4.19, are still functioning
properly, so it is believed the GCE/GS setup is still functioning properly.

The range over which this failed was between 8719 and 8720, the change to
upgrade to 4.20 is the most suspicious CL in this range.
https://crosland.corp.google.com/log/8719.0.0..8720.0.0

Thus this is a long shot, but is the current best known option to try to
recover the tree, if this does not fix it, we can always revert.

BUG=chromium:639314
TEST=./lib/gs_unittest --network

Change-Id: I0ccf71cabd95f25a66a04d7eafb6a7afe121a916
Reviewed-on: https://chromium-review.googlesource.com/374526
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>
(cherry picked from commit 99c0abc81278887d9674dd0cd6aeb5aae783d819)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2348672
Tested-by: Mike Frysinger <vapier@chromium.org>