Update AFDO time out on R59.
Chrome has not updated for more than 14 days on R59, so now the AFDO
mechanism will not allow AFDO profile files to update because it
considers them too old.  This issue is fixed in master/ToT, but
not in R59, so we need to increase the timeout for AFDO on R59 for
now.

BUG=chromium:751209
TEST=None

Change-Id: Ic1ad8f2fecc7e9c48178c8e6433eac0b45b1b505
Reviewed-on: https://chromium-review.googlesource.com/596512
Reviewed-by: Ting-Yuan Huang <laszio@chromium.org>
Tested-by: Grace Kihumba <gkihumba@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>
Trybot-Ready: Caroline Tice <cmtice@chromium.org>