infra/config: Create luci-scheduler task for som-dispatcher.
This is intended to use the luci-scheduler to start the som-dispatcher
builder on the chromeos waterfall 5 minutes after the previous build
finishes via buildbucket.

If this works correctly, we will extend this further to use
luci-scheduler as a replacement for most (all?) builds currently
scheduled via buildbot's scheduler. This is an important step towards
fully moving away from buildbot.

BUG=chromium:751272
TEST=None

Change-Id: Idd4586b6fde847f17468aec1c0e5f4adc68813e2
Reviewed-on: https://chromium-review.googlesource.com/597068
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>