cbuildbot_config: disable unittests for factory configs
The unittests on this ancient branch is failing.  We know some older
versions aren't as hermetic as they should be (and has been addressed
in master), so disable them here.

Practically speaking for factory images, unittests don't have as much
value as we only care whether Chrome runs and can run some stuff.

BUG=chromium:864873
TEST=None

Change-Id: I713128ae46366962c6f923a4b5ae1b3e8a8b057b
Reviewed-on: https://chromium-review.googlesource.com/1146875
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>