paygen: Use gspaths.Build as a variable for all Images [cherry-pick]
Currently, the combination of bucket, version, channel, and board are used in
URIs for Build, Image, Image archive and their accompanying factory
functions (including Payload). However, one can assume that each Image or
Payload is associated with a Build, right? So why not instead of using those
four variables everywhere in the code, add a 'build' key for them and simplify a
lot of things. This allows us to extend the use of Build later for DLCs too
easily. This change shortens the code by more than 150 lines and makes it much
easier to understand and work with.

This also means passing the bucket is mandatory now so we pass the
'chromeos-releases' in the release stage code.

BUG=chromium:862679
TEST=payload-tryjob
TEST=unittests

Change-Id: I86812d45cae8762307aba4aa4d868e2377624e5a
Reviewed-on: https://chromium-review.googlesource.com/c/1483795
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>