chromeos_config: Disable paygen tests in R65.
This is a follow up to CL:1249667 that disabled hwtests, but
missed payload testing on the R65 branch.

Original notes:

The Lakitu LTS branch (R65) does release builds which include "extra"
builders. These builders are running hwtests, which are problematic
for the lab.

Since this branch is otherwise unused, disable all hwtests to avoid
the lab problems. This fix ONLY applies to R65.

BUG=chromium:881376
TEST=chromeos_config_unittest

Change-Id: Id285e80960902464fd4c2075e0f1cc64f617a659
Reviewed-on: https://chromium-review.googlesource.com/c/1308275
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Alex Zamorzaev <zamorzaev@chromium.org>
Reviewed-by: Niranjan Kumar <kumarniranjan@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>