upload_symbols: Refactor to utilize new endpoint
Refactor the symbols upload script to utilize the new authenticated
endpoint.  This will remove the dependency on whitelisting.

Additionally, this is a merge of two CLs; refactor the endpoint as well as remove the dependency on Isolate for duplication processing.

BUG=chromium:912652
TEST='cros_tryjob samus-release-tryjob'

Change-Id: Id6dbf9032c7a497727283f880272ecdae030efdc
Reviewed-on: https://chromium-review.googlesource.com/1559661
Commit-Ready: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1583076
Reviewed-by: David Burger <dburger@chromium.org>
Reviewed-by: Cindy Bayless <cindyb@chromium.org>
Tested-by: Cindy Bayless <cindyb@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>