[chromite] Remove master-lakitu-release
As stated when it was added (CL:1341459), this was the first time it's
been done and apparently it has caused problems at various places in the
system. Turn it down for now until we fix the conflicts between this and
the main, master-release builder.

As planned, this CL should be cherrypicked all the way down to the
oldest stable milestone still supported, i.e., m73.

TEST=config/chromeos_config_unittest
BUG=chromium:959205

Change-Id: I5c74c581e4354ef9d29f286b02f58d9fff46425a
Reviewed-on: https://chromium-review.googlesource.com/1594990
Commit-Ready: Dhanya Ganesh <dhanyaganesh@chromium.org>
Tested-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1602874
Commit-Queue: Ke Wu <mikewu@google.com>
Tested-by: Ke Wu <mikewu@google.com>