build_status: Increase timeout for build start
The timeout for a build to start is 5 minutes.  With timeouts in place
today for the overall build, increasing this timeout to allow for some
queuing on the swarming side is reasonable.  This is primarily due to
release builders causing overruns that have resulted in bot shortages.

BUG=chromium:976707
TEST=`cros_sdk -- /mnt/host/source/chromite/run_tests`

Change-Id: I80207f8f1f22f54bfa70dcf6be9506bc3b2a4c9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1667536
Reviewed-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>
(cherry picked from commit ee08c785ae2217380beef4122ae9fe13e9e20755)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1667923