cbuildbot: Ignore TypeError when making timeline
Currently, these branches are only used for Lakitu builds, and we don't
care about these build timelines. We would rather let this part of the
stage fail and continue, since we do depend on the LinkArtifacts
component. Let's ignore TypeErrors when creating the build timelines
(since that is an error we are seeing right now), and make the root
cause investigation of this issue P2.

BUG=b:140301130
TEST=tryjob

Change-Id: I04a7f06fc648973a4f771f9573866fe53a279c00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1797286
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>