Remove test config override for betty-nyc-arc-release on R78 branch.
Remove the test config override for betty-nyc-arc-release on branch
cros/release-R78-12499.B since this is no longer an enabled
builder. This causes the config-updater to fail as it attempts to roll
in the updated ge_bulid_config.json and there is no site-config
betty-nyc-arc-release for this to match up with. See the attached bug.

Following this commit it also appears that the ge_build_config.json that
goes with this branch, as stored here:

gs://chromeos-build-release-console/build_config.release-R78-12499.B.json

Needs to have a betty-arcmaster section in it as follows:

    {
      "name": "betty-arcmaster",
      "configs": [
        {
          "builder": "RELEASE",
          "experimental": true,
          "leader_board": false,
          "board_group": "",
          "arch": "X86_INTERNAL"
        }
      ]
    },

or a different error will be surfaced after this commit goes through,
during the operation of the config-updater.

TEST=./config/chromeos_config_unittest --update
TEST=./run_tests (with R78 ge_build_config.json and the change noted above)
BUG=chromium:1004843

Change-Id: I28a24631a68b444d9d3df6dbf4ac6359e41dd905
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1810221
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>