chromeos_config update: Skip overrides if config not found.
Skip the application of config overrides when the config is not
found. This can happen in branches after the fact when the
ge_build_config.json is adjusted via GoldenEye/DLM and a config no
longer exists for that release.

This is currently breaking the config-updater, see the attached bug.

BUG=chromium:1083985
TEST=gsutil cp \
  gs://chromeos-build-release-console/build_config.release-R84-13099.B.json \
  config/ge_build_config.json \
  && cros_sdk -- /mnt/host/source/chromite/config/refresh_generated_files

Change-Id: I55a88346566b8e7620ba6736e55e933cdd2c2bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2207392
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>